+++ crates/nostr/src/nips/nip47.rs
@@ -371,10 +371,13 @@ pub struct MakeInvoiceRequest {
     /// Amount in millisatoshis
     pub amount: u64,
     /// Invoice description
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
     /// Invoice description hash
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub description_hash: Option<String>,
     /// Invoice expiry in seconds
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub expiry: Option<u64>,
 }
 
@@ -382,8 +385,10 @@ pub struct MakeInvoiceRequest {
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub struct LookupInvoiceRequest {
     /// Payment hash of invoice
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub payment_hash: Option<String>,
     /// Bolt11 invoice
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub invoice: Option<String>,
 }
 
@@ -399,7 +404,7 @@ pub enum TransactionType {
 }
 
 /// Transaction State
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 pub enum TransactionState {
     /// Pending
     #[serde(rename = "pending")]
@@ -666,16 +671,16 @@ pub struct PayKeysendResponse {
 }
 
 /// Make Invoice Response
+//
+// NOTE: don't add `type`, `state` or `fees_paid`, as they don't make sense here.
+//
 #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
 pub struct MakeInvoiceResponse {
-    /// Transaction type
-    #[serde(rename = "type")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub transaction_type: Option<TransactionType>,
-    /// Transaction state
-    pub state: TransactionState,
-    /// Bolt11 invoice
+    /// Bolt 11 invoice
     pub invoice: String,
+    /// Invoice's payment hash
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub payment_hash: Option<String>,
     /// Invoice's description
     #[serde(skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
@@ -685,20 +690,15 @@ pub struct MakeInvoiceResponse {
     /// Payment preimage
     #[serde(skip_serializing_if = "Option::is_none")]
     pub preimage: Option<String>,
-    /// Invoice's payment hash
-    pub payment_hash: String,
-    /// Amount in millisatoshis
-    pub amount: u64,
-    /// Fees paid in millisatoshis
-    pub fees_paid: u64,
+    /// Amount in msats.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub amount: Option<u64>,
     /// Creation timestamp in seconds since epoch
-    pub created_at: Timestamp,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub created_at: Option<Timestamp>,
     /// Expiration timestamp in seconds since epoch
     #[serde(skip_serializing_if = "Option::is_none")]
     pub expires_at: Option<Timestamp>,
-    /// Optional metadata about the payment
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub metadata: Option<Value>,
 }
 
 /// Lookup Invoice Response
@@ -709,7 +709,8 @@ pub struct LookupInvoiceResponse {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub transaction_type: Option<TransactionType>,
     /// Transaction state
-    pub state: TransactionState,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub state: Option<TransactionState>,
     /// Bolt11 invoice
     #[serde(skip_serializing_if = "Option::is_none")]
     pub invoice: Option<String>,
@@ -880,8 +881,10 @@ pub struct Response {
     /// Request Method
     pub result_type: Method,
     /// NIP47 Error
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub error: Option<NIP47Error>,
     /// NIP47 Result
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub result: Option<ResponseResult>,
 }
 
@@ -1389,7 +1392,8 @@ pub struct PaymentNotification {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub transaction_type: Option<TransactionType>,
     /// Transaction state
-    pub state: TransactionState,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub state: Option<TransactionState>,
     /// Bolt11 invoice
     pub invoice: String,
     /// Invoice's description
@@ -1558,7 +1562,7 @@ mod tests {
             result: Some(ResponseResult::ListTransactions(vec![
                 LookupInvoiceResponse {
                     transaction_type: Some(TransactionType::Incoming),
-                    state: TransactionState::Expired,
+                    state: Some(TransactionState::Expired),
                     invoice: Some(String::from("abcd")),
                     description: Some(String::from("string")),
                     amount: 123,
@@ -1609,7 +1613,7 @@ mod tests {
             Some(ResponseResult::ListTransactions(vec![
                 LookupInvoiceResponse {
                     transaction_type: Some(TransactionType::Incoming),
-                    state: TransactionState::Expired,
+                    state: Some(TransactionState::Expired),
                     invoice: Some(String::from("abcd")),
                     description: Some(String::from("string")),
                     amount: 123,
@@ -1653,7 +1657,7 @@ mod tests {
         );
         let notification_result = NotificationResult::PaymentReceived(PaymentNotification {
             transaction_type: Some(TransactionType::Incoming),
-            state: TransactionState::Settled,
+            state: Some(TransactionState::Settled),
             invoice: String::from("abcd"),
             description: Some(String::from("string1")),
             description_hash: Some(String::from("string2")),